photo_be/requirements.txt
asgiref==3.8.1
boto3==1.34.96
botocore==1.34.96
Django==5.0.4
django-storages==1.14.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
jmespath==1.0.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.1
